@ctumwebaze @austiine04 rapidftr/tracker#52 bunding apk with server
This fix ensures that the apk is downloaded using http rather than https. This is because
if the server is using a self-signed certificate, the android device will not trust the certificate and
will go into an indefinate download loop.

This fix ensure that if a request for an apk comes in over https, it will be redirected to http.