rapidftr/tracker#180 syncing enquiries.
@ctumwebaze @mwangiann
When synchronizing enquiries from the mobile to the web, it was failing due to that fact that the mobile
device was sending along particular fields that are couchdb system fields i.e '_rev' that were causing conflicts.

This fix extends the functionality of the enquiries to allow.
1. creation of a unique identifier for an enquiry before saving it. This is important because the mobile device requires this
field to be present to avoid creation of duplicate records when synchronizing.
2. dealing with history logs of an enquiry during synch.