hide enquiries tab
when enquiries are disabled, the user should not be able to see the
enquiries tab on the home page and any interaction with the enquiries
either by adding, deleting, updating or even listing should result in a
404 error and should display an error message telling the user the page
doesn't exist.

To address the issue above, a system variable was introduced that holds
a value indicating whether enquiries are enabled or not. so in the
application layout file, a conditional check is applied to determine
whether to show the enquiries link based on the value of the system
variable.

A controller filter added as a gate keeper to determine whether any
controller actions should be invoked. the invocation is dependant on the
value of the system variable. And in the event the enquiries are turned
off, the filter stops the execution of the controller action and returns
a 404 http status along with a message to display to the user.