enquiry search type handling
when enquiries are turned off and a user makes a search query as
enquiries as the search type, the application should refuse the search.
To do achieve this, the search controller returns a 400 bad request when
enquiry is used as a search type in the search url.

this fix is also taken further to apply to any search type not supported
by the system for example if a user make of a search type 'boy', the
application will still throw a 400 bad request.